backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
mysql-connector-python
textblob
python-dotenv
